requirements.txt
# Core dependencies
solana>=0.25.0
solders>=0.26.0
click>=8.0.0
pyyaml>=6.0
pycryptodome>=3.17.0
python-dotenv>=1.0.0
base58>=2.1.0
aiohttp>=3.8.0
asyncio>=3.4.3
rich>=13.0.0

# Quantum-hash framework (local)
# Add this line if installing from local path:
# -e ./quantum_hash

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0